feat: PassengerCounter 컴포넌트 분리
